+++ src/main/java/jp/co/sfrontier/ss3/janken_game/model/UserInfo.java
@@ -1,0 +1,35 @@
+package jp.co.sfrontier.ss3.janken_game.model;
+
+import java.io.Serializable;
+/**
+ *ユーザ情報を格納するエンティティクラス
+ */
+public class UserInfo implements Serializable {
+    private int userId;
+    private String mailAddress;
+    private String userName;
+    
+    public int getUserId() {
+        return userId;
+    }
+
+    public void setUserId(int userId) {
+        this.userId = userId;
+    }
+
+    public String getMailAddress() {
+        return mailAddress;
+    }
+
+    public void setMailAddress(String mailAddress) {
+        this.mailAddress = mailAddress;
+    }
+    
+    public String getUserName() {
+        return userName;
+    }
+
+    public void setUserName(String userName) {
+        this.userName = userName;
+    }
+}
