+++ src/main/webapp/WEB-INF/web.xml
@@ -1,67 +1,59 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd"
-    version="3.1" metadata-complete="true">
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd"
+	version="3.1" metadata-complete="true">
 
-    <filter>
-        <filter-name>LoginFilter</filter-name>
-        <filter-class>com.kadai.aws.filter.LoginFilter</filter-class>
-    </filter>
-    <filter-mapping>
-        <filter-name>LoginFilter</filter-name>
-        <url-pattern>/game/Play</url-pattern>
-        <url-pattern>/history</url-pattern>
-    </filter-mapping>
+	<filter>
+		<filter-name>LoginFilter</filter-name>
+		<filter-class>jp.co.sfrontier.ss3.janken_game.filter.LoginFilter</filter-class>
+	</filter>
+	<filter-mapping>
+		<filter-name>LoginFilter</filter-name>
+		<url-pattern>/game/Play</url-pattern>
+		<url-pattern>/history</url-pattern>
+	</filter-mapping>
 
-    <servlet>
-        <servlet-name>LoginServlet</servlet-name>
-        <servlet-class>com.kadai.aws.controller.LoginServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>LoginServlet</servlet-name>
-        <url-pattern>/login</url-pattern>
-    </servlet-mapping>
-    <servlet>
-        <servlet-name>RegisterServlet</servlet-name>
-        <servlet-class>com.kadai.aws.controller.RegisterServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>RegisterServlet</servlet-name>
-        <url-pattern>/register</url-pattern>
-    </servlet-mapping>
-    <servlet>
-        <servlet-name>LogoutServlet</servlet-name>
-        <servlet-class>com.kadai.aws.controller.LogoutServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>LogoutServlet</servlet-name>
-        <url-pattern>/logout</url-pattern>
-    </servlet-mapping>
-    <servlet>
-        <servlet-name>GameServlet</servlet-name>
-        <servlet-class>com.kadai.aws.controller.GameServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>GameServlet</servlet-name>
-        <url-pattern>/game/Play</url-pattern>
-    </servlet-mapping>
-    <servlet>
-        <servlet-name>JankenServlet</servlet-name>
-        <servlet-class>com.kadai.aws.controller.JankenServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>JankenServlet</servlet-name>
-        <url-pattern>/play</url-pattern>
-    </servlet-mapping>
-    <servlet>
-        <servlet-name>HistoryServlet</servlet-name>
-        <servlet-class>com.kadai.aws.controller.HistoryServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>HistoryServlet</servlet-name>
-        <url-pattern>/history</url-pattern>
-    </servlet-mapping>
+	<servlet>
+		<servlet-name>LoginServlet</servlet-name>
+		<servlet-class>jp.co.sfrontier.ss3.janken_game.controller.login.LoginServlet</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>LoginServlet</servlet-name>
+		<url-pattern>/login</url-pattern>
+	</servlet-mapping>
+	<servlet>
+		<servlet-name>RegisterServlet</servlet-name>
+		<servlet-class>jp.co.sfrontier.ss3.janken_game.controller.login.RegisterServlet</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>RegisterServlet</servlet-name>
+		<url-pattern>/register</url-pattern>
+	</servlet-mapping>
+	<servlet>
+		<servlet-name>LogoutServlet</servlet-name>
+		<servlet-class>jp.co.sfrontier.ss3.janken_game.controller.login.LogoutServlet</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>LogoutServlet</servlet-name>
+		<url-pattern>/logout</url-pattern>
+	</servlet-mapping>
+	<servlet>
+		<servlet-name>GameServlet</servlet-name>
+		<servlet-class>jp.co.sfrontier.ss3.janken_game.controller.play.GameServlet</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>GameServlet</servlet-name>
+		<url-pattern>/game/Play</url-pattern>
+	</servlet-mapping>
+	<servlet>
+		<servlet-name>HistoryServlet</servlet-name>
+		<servlet-class>jp.co.sfrontier.ss3.janken_game.controller.history.HistoryServlet</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>HistoryServlet</servlet-name>
+		<url-pattern>/history</url-pattern>
+	</servlet-mapping>
 
 </web-app>
